remove ical, implement gcal into events
