--- src/routes/(site)/events/demo.json/+server.ts
@@ -1,8 +1,0 @@
-import { google } from 'googleapis';
-
-export async function GET() {
-  const events = await listUpcomingEvents();
-  return new Response(JSON.stringify(events), { status: 200 });
-}
-
-
